pyproject.toml
[project]
name = "mcp-esa-server"
version = "0.1.0"
description = "MCP server for esa.io API"
authors = [
    {name = "Naoki Numaguchi", email = "redheru@gmail.com"},
]
dependencies = [
    "requests",
    "python-dotenv",
    "mcp[cli]>=1.7.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.8",
    "httpx>=0.28.1", # Keep httpx for potential future testing needs
    "pytest-mock>=3.14.0",
]

[project.scripts]
mcp-esa-server = "src.mcp_esa_server.server:main"

[tool.ruff]
line-length = 120
target-version = "py313" # プロジェクトで使用するPythonバージョンに合わせてください

[tool.ruff.lint]
select = ["E", "F", "W", "I"] # 基本的なルールセットを選択

[tool.uv]
dev-dependencies = [
    "pytest",
    "ruff",
    "python-dotenv",
    "uv>=0.7.3",
    "pytest-cov>=6.1.1",
]

[tool.pytest.ini_options]
pythonpath = [
    "."
]
markers = [
    "integration: mark test as integration test (requires network and .env)",
]
